refactor: extract the 'analyze' function from index.js
